+++ viteclientts/src/pages/LoginPage.tsx
@@ -10,7 +10,7 @@ import { useTranslation } from 'react-i18next';
 const LoginForm: React.FC = () => {
   const [username, setUsername] = useState<string>('');
   const [password, setPassword] = useState<string>('');
-  const { setPreviousLastLogin, setUserLogin, setHasVisitedBefore, setLogoutInProgress, setTokenAndExpiry, setDeviceId } = useUserStore();
+  const { setUserLogin, setHasVisitedBefore, setLogoutInProgress, setTokenAndExpiry, setDeviceId } = useUserStore();
   const navigate = useNavigate();
   const apiUrl = import.meta.env.VITE_API_URL;
   const { showSnackbar } = useSnackbar();
@@ -32,7 +32,6 @@ const LoginForm: React.FC = () => {
         setTokenAndExpiry(response.data.accessToken);
         setHasVisitedBefore(true);
         setDeviceId(response.data.deviceId);
-        setPreviousLastLogin((response.data.previousLastLogin));
         navigate('/home');
 
         const backendMessageKey = response.data.message;